backend_django/pyproject.toml
[tool.poetry]
name = "pixel-backend"
version = "0.1.0"
description = "Backend Django 5 + DRF pour PixelProwlers"
authors = ["PixelProwlers <contact@pixelprowlers.test>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1"
djangorestframework = "^3.15"
django-cors-headers = "^4.4"
psycopg2-binary = "^2.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
